separate page edit content
