+++ package.json
@@ -1,5 +1,5 @@
 {
-  "version": "11.0.2",
+  "version": "11.1.0",
   "license": "MIT",
   "main": "dist/cjs/index.js",
   "module": "dist/esm/index.js",
