+++ src/components/edit-page/EditPage.tsx
@@ -1,16 +1,14 @@
-import React, { useCallback, useEffect, useMemo, useState } from 'react';
-import { SideNav, SideNavItem } from '../side-nav';
-import { ButtonBox, SideMenuFlexGridItem } from './styles';
+import React from 'react';
+import { ButtonBox } from './styles';
 import { EditPageProps } from './types';
 import { EditPageSkeleton } from './EditPageSkeleton';
 import { useTheme } from '../../providers';
 import { FlexItem } from '../flex-item';
-import { FlexGrid, FlexGridItem } from '../flex-grid';
 import { HeadingSmall } from '../typography';
 import { Button } from '../button';
 import { ButtonKind } from '../../models';
 import { CaretLeftIcon } from '../icons';
-import { Box } from '../box';
+import { EditPageContent } from './EditPageContent';
 
 export const EditPage = ({
   sideNavItems,
@@ -25,67 +23,13 @@ export const EditPage = ({
   cancelText,
   onCancel,
 }: EditPageProps) => {
-  const [tab, setTab] = useState<string>(sideNavItems[0]?.id);
-  const [secondaryTab, setSecondaryTab] = useState<string>();
-
   const {
     theme: {
       current: {
-        sizing: { scale800, scale500, scale1600 },
+        sizing: { scale500 },
       },
     },
   } = useTheme();
-
-  useEffect(() => {
-    if (selectedTab) {
-      setTab(selectedTab);
-    }
-  }, [selectedTab]);
-
-  useEffect(() => {
-    if (selectedSubTab) {
-      setSecondaryTab(selectedSubTab);
-    }
-  }, [selectedSubTab]);
-
-  const getAllSubItems = useCallback(() => {
-    return sideNavItems.reduce((acc: SideNavItem[], item: SideNavItem) => {
-      return [...acc, ...(item.subItems || [])];
-    }, []);
-  }, [sideNavItems]);
-
-  const contentComponent = useMemo(() => {
-    let navItem;
-    const subItems = getAllSubItems();
-
-    if (secondaryTab) {
-      if (subItems.length) {
-        navItem = subItems.find((item) => item.id === secondaryTab && item.secondaryId === tab);
-      } else {
-        navItem = sideNavItems.find((item) => item.id === tab && item.secondaryId === secondaryTab);
-      }
-    } else {
-      navItem = sideNavItems.find(({ id }) => id === tab);
-    }
-
-    if (navItem?.component) {
-      return <>{navItem.component()}</>;
-    }
-
-    return (subItems.length ? [...subItems, ...sideNavItems] : sideNavItems).map(({ id, component: Component }) =>
-      Component ? <Component key={id} /> : null,
-    );
-  }, [sideNavItems, getAllSubItems, tab, secondaryTab]);
-
-  const handleNavigation = (tab: string) => {
-    setTab(tab);
-    setSecondaryTab('');
-  };
-
-  const handleSecondaryTabNavigation = (tab: string, parentId: string) => {
-    setSecondaryTab(tab);
-    setTab(parentId);
-  };
 
   const onCancelClick = () => {
     if (onCancel) {
@@ -120,20 +64,7 @@ export const EditPage = ({
           )}
         </ButtonBox>
       </FlexItem>
-      <FlexGrid flexGridColumnCount={2} flexGridColumnGap={scale800} flexGridRowGap={scale1600}>
-        <SideMenuFlexGridItem>
-          <Box>
-            <SideNav
-              items={sideNavItems}
-              activeItemId={tab}
-              activeSecondaryItemId={secondaryTab}
-              handleSecondaryTabNavigation={handleSecondaryTabNavigation}
-              handleNavigation={handleNavigation}
-            />
-          </Box>
-        </SideMenuFlexGridItem>
-        <FlexGridItem>{contentComponent}</FlexGridItem>
-      </FlexGrid>
+      <EditPageContent sideNavItems={sideNavItems} selectedTab={selectedTab} selectedSubTab={selectedSubTab} />
     </>
   );
 };
