+++ src/components/edit-page/index.ts
@@ -1,6 +1,7 @@
 export * from './EditPage';
 export * from './styles';
 export * from './EditPageContainer';
+export * from './EditPageContent';
 export * from './EditPageSkeleton';
 export * from './SideMenuSkeleton';
 export * from './types';
